Fixes #348
The problem was that after a version of FSharp.Data.dll was loaded, it's not possible to load another one. When we had only one profile it was fine, because full .Net was loaded normally, and the profile was loaded with reflection only, so there was no clash.
I turned all framework version to reflection only load, and gave different version numbers to all assemblies to avoid this problem.