Adds HTTP Basic Auth and Digest support
Currently the type providers do not support authorization.
This patch adds support for interpreting the userinfo part of the
provided URL, constructing HTTP Auth and Digest authorization
credentials (as SecureString), creating and passing them to
credentials cache and adding it to the web request also constructed.
The credentials are removed from the URL both before building the
credentials and the request. In presence of authorization credentials
the GetSample methods are disabled in order to divulging the credentials
to the generated assembly unbeknownst to the user.

Fixes #158