Fixes an issue where passwords are not transmitted as UTF-8 Base64 encoded strings, but rather encoded using the system codepage in use at the encoding moment. This could result in situation where non-[A-Za-z] passwords won't work on some locales. The fix is done by implementing a IAuthenticationModule. It looks like this is the approach to take to introduce a OAuth2 authentication too, with either implementing ICredential or inheriting from NetworkCredential appropriately.
Also, it seems like an approach where the user/programmer defines which credentials should be added to the cache should be preferred. It could be in that case the type provider interface should expose a method to add directly credentials either implementing ICredential or inheriting from NetworkCredential.

Also, code reorganized.

#fixes 158