Add HTTP Basic Authentication support for portable libraries
Issues:
- Portable libraries do not support CredentialCache and thus
multiple credentials, which may be necessary in case of
proxies and credentials.

- Portable libraries do not support PreAuthenticate.

- Portable libraries do not support redirect follow-ups.
Redirection has been defined for full profile, and it will affect
all and not just authentication scenarios.

- It's unclear if authentication function properly in Silverlight
using the browser stack or should "WebRequest.RegisterPrefix
("http://", WebRequestCreator.ClientHttp)
WebRequest.RegisterPrefix("https://",
WebRequestCreator.ClientHttp)" added.

In that case also namespaces
System.Windows.Browser and System.Net.Browser should
be opened and appropriate libraries added.

Some of these issues are concerned if one wants to handle authentication
in presence of redirects and how to build the case of OAuth.