Add EmbeddedResource parameter to Json and Xml providers
This was triggered by the issue in Twitter: https://github.com/fsprojects/FSharp.Data.Toolbox/issues/1

The problem is that `FSharp.Data.Toolbox.Twitter.dll` exposes the types
generated by the JSON type provider to the caller - so the methods in the
Twitter API return types generated by the type provider from sample JSON
files (I think this is reasonable - if we had to re-write all the type
declarations by hand, then this would be hard work).

The problem is that the samples are in a local folder (`json/foo.json`)
and when you build `FSharp.Data.Toolbox.Twitter.dll`, you do not want to
always keep them around (because referencing the Twitter API will trigger
the type provider and the type provider will be looking for them).
Embedding the samples inline also does not work - they are just too long
and the F# compiler cannot deal with them.

With this change, we can say `JsonProvider<"sample.json", EmbeddedResource="MyLib,sample.json">`.
The type provider will first try to load `MyLib` and find the resource
there (so, when you are referencing compiled `FSharp.Data.Toolbox.Twitter.dll`,
it will find the resource there). If this fails, it falls back to standard
directory search (which handles the case when you are compiling
`FSharp.Data.Toolbox.Twitter.dll` itself).

I'm not entirely happy with this solution (feels a bit hacky), but I do not
think that there is any other way - aside from making the type providers
generative, which would be nice thing to do, but AFAIK not something that'll
happen very soon.