Accept more MIME types in HTTP requests
Currently when reading text from a URL, the MIME type is set to the
specific format being requested, e.g. application/xml for XML. This
will cause the request to fail with HTTP error 406 if the resource
has a different MIME type, such as text/plain or application/atom+xml.
This commit modifies it to prefer specific types, followed by */* so
that any MIME type is accepted.