Updates to fix for issue #688
Added logic to look ahead at the second line and cache it when it
exists.  This eliminates the need for a new reader.  Changed the test
for the empty trailing header to use String.IsNullOrWhiteSpace instead
of checking for zero length.  Also eliminate the update to separators
since a new reader is no longer created.